requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.20
pandas==2.1.1
requests==2.31.0
